Give place/ a consistent shape and move transit UI to the transit module
place/ held the two largest transit components (a stop departures card and a
full stop board) plus a GTFS route panel misleadingly named RouteDetailPage,
while transit/ held only small leaf primitives. Those move to
transit/{departures,routes}/ under names that say what they are, and the two
visit-history components move to timeline/.

Within place/, the four one-file folders and the seven loose files collapse into
card/, details/, header/ and widgets/ beside their only consumers. Place.vue
becomes PlacePanel.vue so it stops shadowing the Place type it imports.
