+++ web/src/components/place/card/PlaceCategoryIcon.vue
