+++ web/src/components/place/card/PlaceList.vue
