+++ web/src/components/place/card/PlaceListItem.vue
