+++ web/src/components/place/details/NearbyCategories.vue
