+++ web/src/components/place/details/PlaceDisplayChips.vue
