+++ web/src/components/place/details/PlaceTypeChip.vue
