+++ web/src/components/place/widgets/PlaceBikeshare.vue
