+++ web/src/components/transit/routes/TransitRouteDetailPanel.vue
