+++ web/src/views/search/Search.vue
@@ -12,7 +12,7 @@ import type { MapCamera, MapBounds } from '@/types/map.types'
 import { useSearchService, type BrandHeader } from '@/services/search.service'
 import type { Place } from '@/types/place.types'
 import type { SearchResult } from '@/types/search.types'
-import PlaceList from '@/components/place/PlaceList.vue'
+import PlaceList from '@/components/place/card/PlaceList.vue'
 import FilterChips from '@/components/map/FilterChips.vue'
 import ErrorMessage from '@/components/ErrorMessage.vue'
 import { useRouter } from 'vue-router'
